requirements.txt
Django
gunicorn
django-heroku
django-bootstrap-v5
django-allauth
dj-database-url
dj-static
psycopg2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django-googledrive-storage
mock
numpy
opencv-python
django-crispy-forms
django-storages[google]
google.cloud